Sniff only the filters object, not the surrounding json value.
So, it looks like summarize has _optional_ filters. Perhaps search does
too? At any rate, in summarize we're building the full body for the
caller rather than letting them provide a partial body with a filters
key in it.

Body building in this lib is wonky and we will ultimately
need to conform, likely whenever we switch to using the builder pattern
for these requests.

For now, I'm just doing the wonky thing that happens to make sense this
time around which requires us to adjust the mime sniffing to not
anticipate the outer json structure.