Adds a `gzip` feature to enable reqwest's feature of the same name.
I wrote a small program that will fetch the first 2,000 tasks from
shotgun with the field list set to "*" (meaning we'd be fetching all
425+ fields per task). The program tracks the elapsed time of the
awaited `search()` call.

Without gzip enabled

```
70.873702357s
67.787549573s
67.88784679s
69.171983736s
68.952043381s
68.800091773s
```

With gzip enabled

```
93.690752643s
62.35949705s
111.377153161s
96.378771547s
67.907775058s
69.08898674s
```

I'm not quite sure what to make of the results. This work was prompted
by users posting to the reqwest issue board suggesting they saw a
dramatic slowdown (not related to tokio) when they upgraded from reqwest
0.9 to 0.10, and they noticed the gzip feature was previously on by
default, but no longer.

Usually, when you look at gzip for http traffic, you don't necessarily
see a *speedup* per se. Usually the timings are similar to those without
compression enabled, but you trade time "on the wire" for time spent
decompressing once the data has been received by the client. In this,
it's not surprising to find the gzip-enabled timings to be mostly
similar.

I can't explain why there was *more variance* in the times with
the gzip feature enabled, but I can't be certain it wasn't just
"atmospheric conditions" over on the staging server. Stil, the build
without the gzip feature enabled was remarkably consistent...

With all this, I think we should include this feature but leave it off by
default, adoping the same policy as reqwest itself.