SGRS-22 Add some wiremock tests.
This diff aimed to provide a working example of how we might use
wiremock to expand our test coverage.

A tricky thing here was deciding what to test. I chose to focus on
probing the error handling code, using a user login as the basis.

For the most part, this lib has no business testing things like the data
returned by certain filters, but we can probe the machinery in terms of
how it decides between Ok and Err. We should, however aim to expand this
to include decode tests for responses *we* provide structs for (such as
`TokenResponse` and `ErrorObject`, etc. We can add more mocked responses
for new cases as they are seen in the wild.