Add file uploads.
Flows for both Shotgun and S3 storage services, with multipart uploads (only
when using S3) are in place.

There is, however, and outstanding issue involving multipart uploads.
The Shotgun REST API currently responds 404 when attempting to use
multipart on a record-based upload (as opposed to specifying a field).

Shotgun has filed the bug as SG-20292, and discussion of the bug can be
found at the url below (assuming you have access to the support forum):

- <https://support.shotgunsoftware.com/hc/en-us/requests/117070>