refs SGRS-25 Add builder for `Shotgun::entity_relationship_read()`.
Yet another builder introduced to give callers the chance to ignore the
(seldom-used in my experience) retired/active/archived options.

It's not lost on me that this is a significant code addition just to
make up for the fact rust doesn't support optional parameters. Builders
don't sting as much when there are many optional params in play, but for
this it feels a little punitive.

Eventually these modules I'm spawning as I extract these builders out of
methods in `src/lib.rs` will "fill out" with tests/docs making it more
worth it.