refs SGRS-25 Summarize callsite ergonomics.
Adds `From` impls for the various intermediate structs used to call
`Shotgun::summarize()`.

The aim here is to try and give callers a shortcut for convenience, so
tuples with flexible trait-based elements are coerced into the final
structs via thesee `From`s.

Additionally, the return from summarize's new builder is no longer generic.
Instead it returns a `SummarizeResponse` which models the recursive
structure as far as it can, leaving `serde_json::Value`s at the leaf
parts where the summaries are actually calculated.