refs SGRS-25 callsite ergonomics for schema-ish stuff.
There are several mostly related methods/symbols relating to
schema and/or schema fields that felt mostly like they could live
together in a `schema` module. They have some overlap with their types,
probably because they deal with CRUD ops of a single thing or whatever.

Anyway, this adds some new `From` impls to help build `CreateUpdateFieldProperty`.

It also takes the (2 field) Request- structs we had been accepting at
the callsite and breaks those fields up into separate params (only to
assemble them inside the function body).

I avoided adding a builder for any of this. We had one place where we
accept an Option<i32> for a project id and I think it's low enough
friction to just pass a `None` in there. I suspect this is better
signaling too since certain fields require the project to be specified -
better to put the fact a project id is sometimes expected in the
caller's face.