Add 'type-safe' filter API.
A previous attempt with traits ended poorly.
Putting emphasis on enums across the board this time.

Complex filters now work, technically, but there are some warts with how
you have to be careful to convert Filter -> ComplexFilter in situations
where you are using both together.

Not sure if the solution to the warts is to add careful documentation,
or to go the extra mile to find a more flexible approach types-wise.