Treat examples in README.md as doctests (#12)
* Add doc-comment to build deps. Treat README.md as doctest.

The doc-comment crate provides macros for building doctests from
external files (like a readme!)

This rev adds the dep and a mod, conditionally compiled when cargo
builds doctests, that will run code blocks in the readme as if they were
doc comments in the library source.

These new tests are currently broken! Fixes coming in the next rev.

* Fixup examples in readme.

Now that the readme code blocks are treated as doctests, we need to
actually make sure they compile!

The blocks that aren't rust need to cite some other lang at the top of
the fence. CLI examples are marked as `text`, for example.