Re-export the entire reqwest crate.
There are many types involved in configuring a reqwest client, so rather
than try to cherry-pick the stuff we think will be useful, it might be
better to just re-export the whole thing.

This diff does the re-export under a `transport` module, which could be
useful in the future if we end up offering support for different http clients.

In that situation, there'd be a lot of conditional compilation/feature
checking so we'd want the option to move all that noise into a separate
file.