Change select references to "Shotgun" to instead use "ShotGrid."
Autodesk have rebranded the product from _Shotgun_ to _ShotGrid_. To
reflect this, many objects and comments needed to be updated.

Not that I anticipate Autodesk will rename the product again (not in the
near-term anyway), the client object itself has been renamed from
`Shotgun` to `Client`, which in turn prompted a name change for the
`with_client()` method. This is now the `Client::with_transport()`
method (following the precedent set by the addition of the `transport`
module in #7).

Some notable exceptions to this, preventing an outright find/replace,
are URLs to the support site, mentions of old API details in the
Changelog and the custom mime-types used to represent filter data (which
seem to be keeping the old "shotgun" nomenclature.

Fixes #4.